third_party/WebKit/LayoutTests/platform/win7/virtual/outofblink-cors-http/http/tests/eventsource/workers/eventsource-cors-redirect-expected.txt

[Worker] Test that basic EventSource cross-origin requests fail on redirect.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ../script-tests/eventsource-cors-redirect.js
PASS [Worker] es = new EventSource("/resources/redirect.php?code=307&url=http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php") did not throw exception.
FAIL [Worker] es.readyState should be 2. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
